refactor: asset → assets로 경로 변경 및 import 경로 수정
