+++ src/components/post-comment-input.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image'
-import defaultProfile from '~/assets/svg/default-profile.svg'
+import defaultProfile from '~/assets/svgs/default-profile.svg'
 import { Button } from '~/components/ui/button'
 import { Input } from '~/components/ui/input'
 import { Send } from 'lucide-react'
